fix: check exist add to cart and remove from cart demarket

+++ src/controllers/demarket/Cart.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from "express";
 import { StatusCodes } from "http-status-codes";
-import { InternalServerError, NotFound } from "../../errors";
+import { BadRequest, InternalServerError, NotFound } from "../../errors";
 import accountService from "../../services/demarket/Account.service";
 import nftService from "../../services/demarket/Nft.service";
 import cartService from "../../services/demarket/Cart.service";
@@ -8,7 +8,7 @@ import cartService from "../../services/demarket/Cart.service";
 class CartController {
     /**
      * @method POST => DONE
-     * @description Add to cart
+     * @description ADD TO CART
      * @param request
      * @param response
      */
@@ -22,8 +22,12 @@ class CartController {
             const cartExist = await cartService.findCartByAccountId(String(accountId));
             if (!cartExist) {
                 const newCart = await cartService.createCartByAccountId(String(accountId));
+                const nftExistCart = await cartService.findNftExistAccount({ nftId: String(nftId), cartId: String(newCart.id) });
+                if (nftExistCart) return response.status(StatusCodes.BAD_REQUEST).json(new BadRequest("Nft exist in cart"));
                 await cartService.addNftToCart({ nftId: String(nftId), cartId: newCart.id });
             } else {
+                const nftExistCart = await cartService.findNftExistAccount({ nftId: String(nftId), cartId: String(cartExist.id) });
+                if (nftExistCart) return response.status(StatusCodes.BAD_REQUEST).json(new BadRequest("Nft exist in cart"));
                 await cartService.addNftToCart({ nftId: String(nftId), cartId: cartExist.id });
             }
             response.status(StatusCodes.OK).json({ message: "Add nft to cart successfully" });
@@ -34,7 +38,7 @@ class CartController {
 
     /**
      * @method DELETE => DONE
-     * @description Remove from cart
+     * @description REMOVE FROM CART
      * @param request
      * @param response
      */
@@ -49,8 +53,12 @@ class CartController {
             const cartExist = await cartService.findCartByAccountId(String(accountId));
             if (!cartExist) {
                 const newCart = await cartService.createCartByAccountId(String(accountId));
+                const nftExistCart = await cartService.findNftExistAccount({ nftId: String(nftId), cartId: String(newCart.id) });
+                if (nftExistCart) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Nft not exist in cart"));
                 await cartService.removeNftFromCart({ nftId: String(nftId), cartId: newCart.id });
             } else {
+                const nftExistCart = await cartService.findNftExistAccount({ nftId: String(nftId), cartId: String(cartExist.id) });
+                if (!nftExistCart) return response.status(StatusCodes.NOT_FOUND).json(new NotFound("Nft exist in cart"));
                 await cartService.removeNftFromCart({ nftId: String(nftId), cartId: cartExist.id });
             }
             response.status(StatusCodes.OK).json({ message: "Remove nft from cart successfully" });
